publish

+++ publish.sh
@@ -14,8 +14,11 @@ fi
 
 # emacs in batch mode doesn't include font-locking, so everything is broken.
 # elisp -b "(progn (ns/core) (ns/extra) (ns/development) (ns/style) (ns/blog-generate))"
-elisp -w "(progn (ns/core) (ns/extra) (ns/development) (ns/style) (ns/blog-generate))"
-# elisp_timeout=120 elisp "(ns/blog-generate)"
+# elisp -w "(progn (ns/core) (ns/extra) (ns/development) (ns/style) (ns/blog-generate))"
+if ! elisp_timeout=120 elisp "(ns/blog-generate)"; then
+    echo "generation failure"
+    exit 1
+fi
 
 rm -rf /tmp/notes.neeasade.net
 cp -r site /tmp/notes.neeasade.net
@@ -27,7 +30,7 @@ git checkout master
 git fetch origin
 git reset --hard origin/master
 
-rm -rf *
+# rm -rf *
 cp -r /tmp/notes.neeasade.net/* ./
 
 git add --all
